Refactor IR sensor array modules
