+++ src/heart/peripheral/ir_sensor_array/constants.py
@@ -1,0 +1,3 @@
+SPEED_OF_LIGHT = 299_792_458.0  # metres per second
+DEFAULT_DMA_BUFFER_SIZE = 64
+DEFAULT_RADIAL_LAYOUT_RADIUS = 0.12
